updates folder structure
