+++ src/components/Button/Button.stories.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Button from './Button';
+import Button from '.';
 
 export default {
   title: 'Button',
