+++ src/components/Firebase/index.js
@@ -1,0 +1,5 @@
+import FirebaseContext, { withFirebase } from './FirebaseContext';
+import Firebase from './firebase';
+
+export default Firebase;
+export { FirebaseContext, withFirebase };
