+++ src/components/Footer/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Container from '../Container/Container';
+import Container from '../Container';
 
 
 export default function Footer() {
