+++ src/components/Form/FormInput/FormInput.stories.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import FormInput from './FormInput';
+import FormInput from '.';
 
 export default {
   title: 'Form/FormInput',
