+++ src/components/Header/Header.stories.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import Header from './Header';
+import Header from '.';
 
 const store = {
   getState: () => {
