+++ src/components/Modal/index.js
@@ -1,0 +1,5 @@
+import ModalContainer from './ModalContainer';
+import ModalContent from './ModalContent';
+import MODAL_IDS from './modalIds';
+
+export { ModalContainer, ModalContent, MODAL_IDS };
