+++ src/components/profile/ProfileNav.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import List from '../List/List';
-import NavItem from '../NavItem/NavItem';
+import List from '../List';
+import NavItem from '../NavItem';
 import profileNavItems from '../../content/profileNav';
