+++ src/components/profile/ProfileTabs/tabs/index.js
@@ -1,0 +1,15 @@
+import Connections from './Connections';
+import Discussions from './Discussions';
+import Profile from './Profile';
+import Reviews from './Reviews';
+import Watched from './Watched';
+import Watchlist from './Watchlist';
+
+export {
+  Connections,
+  Discussions,
+  Profile,
+  Reviews,
+  Watched,
+  Watchlist
+};
