+++ src/components/profile/index.js
@@ -1,0 +1,5 @@
+import ProfileNav from './ProfileNav';
+import ProfileHeader from './ProfileHeader';
+import ProfileTabs from './ProfileTabs';
+
+export { ProfileNav, ProfileHeader, ProfileTabs };
