+++ src/screens/HomeScreen/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import Hero from '../../components/Hero/Hero';
-import Features from '../../components/Features/Features';
+import Hero from '../../components/Hero';
+import Features from '../../components/Features';
 
 
 export default function HomeScreen() {
