+++ src/screens/ResetPasswordScreen/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Container from '../../components/Container/Container';
+import Container from '../../components/Container';
 import { ResetPasswordForm } from '../../components/Auth/ResetPassword';
