+++ src/storybook.test.js
@@ -1,5 +1,5 @@
 import initStoryshots from '@storybook/addon-storyshots';
 
 initStoryshots({
-  storyKindRegex: /^((?!Modal).)*$/
+  storyKindRegex: /^((?!ModalContent).)*$/
 });
